+++ CavaService.hpp
@@ -12,12 +12,12 @@ public:
     CavaService();
     ~CavaService();
 
-    void enqueue(double *data);
+    void enqueue(float *data);
     void process_queue();
 
-    sigc::signal<void(double*)> signal_data;
+    sigc::signal<void(float*)> signal_data;
 private:
-    std::queue<double*> queue;
+    std::queue<float*> queue;
     std::mutex queue_mutex;
 
     Glib::Dispatcher dispatcher;
